compose/k2o/requirements.txt
fastapi==0.47.1  # Ensure fastapi is up-to-date
faker
requests
pyjwt[crypto]
